refactoring enzo workflow to parallel other front end workflows
